Remove misleading @auth statements
The authentication required for Respoke actions is nuanced and not
really suited to the simple solution of using @auth in the jsdoc.
Most any action that a normal user ('app-token' authentication) can
perform, an admin can also perform either over websocket or the rest
api.

This patch removes the @auth statements to prevent confusion. We'll
need to just come up with a better way of indicating the
authentication required for each action.