Clean up connect errors
The client was relaying connect_error and connect_timeout socket.io
events, neither of which are events in our version of socket.io.

There is a connect_failed event, but most connections failures are
actually emitted as error events. For consistency, we'll emit
connect_failed as an error through our API.