Remove autoreconnect feature. It is handled by socket.io.
Web socket connections were leaking when auth type was App-Secret. This is
because we added autoreconnect code and it duplicated what socket.io was
doing. I am unsure why this was not clear at the time it was implemented.
Nonetheless, we fall back to using socket.io's reconnection logic, but
only when the web socket was initiated using an App-Secret (that token
lives forever).

Because endpoint brokered auth might work sometimes but fail most of the
time due to the App-Token expiring, we explicitly forbid App-Token web
sockets from using socket.io's reconnect logic. So they will not reconnect
at all - it is an implementation detail left to the user of the library.
They would have needed to obtain an auth token for the initial connection
anyways, so presumably on disconnect they would repeat that process - like
one would do with respoke.js for the browser.

fix: MER-4681

Consider a major version bump due to different behavior for brokered
endpoint auth clients. On the other hand, it is fixing definitely-broken
behavior, so maybe this can be am minor bump.